+++ src/main/java/de/dm/intellij/liferay/schema/LiferayNpmBundlerRcJsonSchemaFileProvider.java
@@ -1,0 +1,52 @@
+package de.dm.intellij.liferay.schema;
+
+import com.intellij.openapi.vfs.VfsUtil;
+import com.intellij.openapi.vfs.VirtualFile;
+import com.jetbrains.jsonSchema.extension.JsonSchemaFileProvider;
+import com.jetbrains.jsonSchema.extension.SchemaType;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.net.URL;
+
+public class LiferayNpmBundlerRcJsonSchemaFileProvider implements JsonSchemaFileProvider {
+
+	@Nullable
+	private final VirtualFile schemaFile;
+
+	public LiferayNpmBundlerRcJsonSchemaFileProvider() {
+		this.schemaFile = getResourceFile();
+	}
+
+	private static VirtualFile getResourceFile() {
+		URL url = LiferaySettingsJsonSchemaFileProvider.class.getResource("/com/liferay/schema/npmbundlerrc-schema.json");
+		if (url != null) {
+			return VfsUtil.findFileByURL(url);
+		}
+		return null;
+	}
+
+	@Override
+	public boolean isAvailable(@NotNull VirtualFile file) {
+		return file.isValid() && ".npmbundlerrc".equals(file.getName());
+	}
+
+	@NotNull
+	@Override
+	public String getName() {
+		return "Liferay NPM Bundler Configuration";
+	}
+
+	@Nullable
+	@Override
+	public VirtualFile getSchemaFile() {
+		return schemaFile;
+	}
+
+	@NotNull
+	@Override
+	public SchemaType getSchemaType() {
+		return SchemaType.embeddedSchema;
+	}
+
+}